+++ app/src/main/java/com/techcourse/dao/UserDao.java
@@ -1,6 +1,8 @@
 package com.techcourse.dao;
 
 import com.techcourse.domain.User;
+import nextstep.jdbc.JdbcTemplate;
+import nextstep.jdbc.RowMapper;
 
 import javax.sql.DataSource;
 import java.util.List;
