+++ jdbc/src/main/java/nextstep/jdbc/PreparedStatementSetter.java
@@ -1,4 +1,4 @@
-package com.techcourse.dao;
+package nextstep.jdbc;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
